Move workflow DueDate component to common ♻️
